requirements.txt
torch==1.8.1
numpy==1.19.2
torchvision==0.9.1
opencv_python==4.5.1.48
matplotlib==3.3.2
imageio==2.9.0
Pillow==8.2.0
